Added employee management API
Added an API that lets you add or delete employees from the database. Returns all employees currently in database unless there is an error deleting. If an employee is not found when a delete request is made it returns a single employee with the request type of "Employee not found" and all other attributes null.