Tiny housekeeping update
Changed the name of a folder to avoid confusion. Found and fixed an error that cropped up when I updated the employee.cs file and neglected to update a call in the employeemanagementAPI. Will reupload to amazon and double check that the format of the database conforms to required standards.