Fixed the employee management api. Should be production ready
Employee management is fixed. Should be production. Employee class is fully fleshed out and up to spec with latest database changes. Working on login API next then the punch API. The punch class will need to be revamped. Added in some in-code documentation and will be working on the wiki soon.