Fixed a couple things
Changed to .net 4.7.1 because it was on 4 for some reason and so syste.net.http was not showing up. Fixed a small data-retrieval thing with punchRecord that was causing extra read requests. frmLogin now will log you in or back from lunch if your last punch was out or lunchout. Haven't made it pull up the clock form for other cases due to the fact that I forgot how.