CSV format changed and small bugfixes
Changed how the CSV is formatted. As it is a report that will be mainly read by human eyes, formatting is paramount. It now displays the employee name instead of empID and puts the punchDate right next to the punchTime. I also added a small method to the API that will retrieve an employee name when given an empID. Added a bunch of comments in the main program to help with future development and readability.