Hardened admin form against culture discrepancies
Found a problem with the admin form regarding culture and the DateTime object. Due to the Date and Time on a different computer than the primary development machine, DateTime objects were getting pulled out with dashes instead of slashes. Added some System.Globilization options to pull all dates as InvariantCulture to ensure proper formatting throughout. This fixed the error. Note: This was not an issue for anything that was processed serverside.